# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "fiber-layout 接口文档",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/args/{arg}": {
            "get": {
                "description": "返回路劲参数",
                "tags": [
                    "fiber-layout"
                ],
                "summary": "path arg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "arg",
                        "name": "arg",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/bind/{arg}": {
            "post": {
                "description": "绑定uri和body参数，并进行参数校验",
                "tags": [
                    "fiber-layout"
                ],
                "summary": "bind json test",
                "parameters": [
                    {
                        "type": "string",
                        "description": "arg",
                        "name": "arg",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}